requirements.txt
telethon
python-dotenv
google-generativeai